requirements.txt
asgiref
cffi
click
cryptography
h11
pycparser
pydantic
PyJWT
starlette
uvicorn
fastapi
requests
psycopg2
Pillow
reportlab
google-api-python-client
google-auth-httplib2
google-auth-oauthlib